refactor: Changed the package structure
