+++ src/main/java/qna/domain/history/delete/DeleteHistoryRepository.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.history.delete;
 
 import org.springframework.data.repository.CrudRepository;
