+++ src/main/java/qna/domain/type/ContentType.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.type;
 
 public enum ContentType {
     QUESTION, ANSWER;
