+++ src/main/java/qna/service/DeleteHistoryService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import qna.domain.DeleteHistory;
-import qna.domain.DeleteHistoryRepository;
+import qna.domain.history.delete.DeleteHistory;
+import qna.domain.history.delete.DeleteHistoryRepository;
 
 import java.util.List;
