+++ src/test/java/qna/domain/user/UserTest.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.user;
 
 public class UserTest {
     public static final User JAVAJIGI = new User(1L, "javajigi", "password", "name", "javajigi@slipp.net");
