+++ src/test/java/qna/service/QnaServiceTest.java
@@ -6,8 +6,14 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import qna.domain.answer.Answer;
+import qna.domain.history.delete.DeleteHistory;
+import qna.domain.question.Question;
+import qna.domain.question.QuestionRepository;
+import qna.domain.question.QuestionTest;
+import qna.domain.type.ContentType;
+import qna.domain.user.UserTest;
 import qna.exception.CannotDeleteException;
-import qna.domain.*;
 
 import java.time.LocalDateTime;
 import java.util.Arrays;
